Improve context isolation
This patch removes the use of @electron/remote and enables contextIsolation. All communication between the main and
renderer processes, mostly involving spell checking state, now happens over defined IPC channels.
