+++ main.ts
@@ -1,11 +1,8 @@
-import { app, BrowserWindow, Menu, protocol, shell } from 'electron';
+import { app, BrowserWindow, dialog, ipcMain, Menu, MenuItem, protocol, shell } from 'electron';
 import * as path from 'path';
 import * as url from 'url';
-
-// https://www.electronjs.org/docs/breaking-changes#removed-remote-module
-// https://github.com/electron/remote/blob/main/docs/migration-2.md
-import { initialize, enable } from '@electron/remote/main';
-initialize();
+import ContextMenu from 'electron-context-menu';
+import { getDicts } from './dicts';
 
 const IS_DEV = process.argv.slice(2).includes('--is-dev');
 
@@ -35,7 +32,7 @@ function createWindow() {
 
 			allowRunningInsecureContent: false,
 			nodeIntegration: false,
-			contextIsolation: false, // Despite what the Electron docs say, this now blocks access to window, so I have to disable it
+			contextIsolation: true,
 			spellcheck: false, // We handle spellcheck with custom code in `preload.ts`
 			preload: preloadPath
 		}
@@ -43,44 +40,54 @@ function createWindow() {
 
 	window.webContents.session.setSpellCheckerEnabled(false);
 
-	// Enable remote
-	enable(window.webContents);
-
-	const appMenu = Menu.buildFromTemplate([
-		{
-			label: (process.platform === 'darwin') ? app.getName() : 'File',
-			submenu: [
-				{ role: 'quit' }
-			]
-		},
-		{
-			id: 'edit-menu',
-			label: 'Edit',
-			submenu: [
-				{ role: 'undo' },
-				{ role: 'redo' },
-				{ type: 'separator' },
-
-				{ role: 'copy' },
-				{ role: 'cut' },
-				{ role: 'paste' },
-				{ role: 'selectAll' },
-				{ role: 'delete' },
-				{ type: 'separator' }
-			]
-		},
-		{
-			label: 'View',
-			submenu: [
-				{ role: 'zoomIn' },
-				{ role: 'zoomOut' },
-				{ role: 'resetZoom' }
-			]
-		},
-		{ role: 'windowMenu' }
-	]);
-
-	Menu.setApplicationMenu(appMenu);
+	ipcMain.on('initalShouldSpellCheck', (_event, shouldSpellCheck) => {
+		const appMenu = Menu.buildFromTemplate([
+			{
+				label: (process.platform === 'darwin') ? app.getName() : 'File',
+				submenu: [
+					{ role: 'quit' }
+				]
+			},
+			{
+				id: 'edit-menu',
+				label: 'Edit',
+				submenu: [
+					{ role: 'undo' },
+					{ role: 'redo' },
+					{ type: 'separator' },
+		
+					{ role: 'copy' },
+					{ role: 'cut' },
+					{ role: 'paste' },
+					{ role: 'selectAll' },
+					{ role: 'delete' },
+					{
+						type: 'checkbox',
+						label: 'Spell Checking',
+						checked: shouldSpellCheck,
+						click: async menuItem => {
+							window?.webContents.send('updateShouldSpellCheck', menuItem.checked);
+							await dialog.showMessageBox({ message: `When you restart µPad, the spell checker will be ${menuItem.checked ? 'enabled' : 'disabled'}.` });
+						}
+					},
+					{ type: 'separator' }
+				]
+			},
+			{
+				label: 'View',
+				submenu: [
+					{ role: 'zoomIn' },
+					{ role: 'zoomOut' },
+					{ role: 'resetZoom' }
+				]
+			},
+			{ role: 'windowMenu' }
+		]);
+		Menu.setApplicationMenu(appMenu);
+	
+		initSpellcheck(shouldSpellCheck)
+			.catch(e => console.error(e));
+	});
 
 	window.loadURL(url.format({
 		pathname: 'index.html',
@@ -107,6 +114,61 @@ function createWindow() {
 function quitApp() {
 	window = null;
 	app.quit();
+}
+
+async function initSpellcheck(shouldSpellCheck: boolean): Promise<void> {
+	if (!shouldSpellCheck) {
+		ContextMenu({ showInspectElement: false });
+		return;
+	}
+
+	const dicts = await getDicts();
+
+	(() => {
+		ContextMenu({
+			menu: (defaultActions, params) => {
+				const spellSuggestions = params.misspelledWord
+					? Array.from(new Set([...dicts.AU.suggest(params.misspelledWord), ...dicts.US.suggest(params.misspelledWord)]))
+					: [];
+
+				const spellSuggestionOptions = spellSuggestions.map(w => {
+					return {
+						label: w,
+						click(selected) {
+							window?.webContents.replaceMisspelling(selected.label)
+						}
+					};
+				});
+
+				const learnSpelling = params.misspelledWord ? [
+					defaultActions.separator(),
+					{
+						label: 'Add to dictionary',
+						click() {
+							window?.webContents.send('addToUserDict', params.misspelledWord);
+						}
+					}
+				] : [];
+
+				return [
+					...spellSuggestionOptions,
+					...learnSpelling,
+					defaultActions.separator(),
+					defaultActions.lookUpSelection({}),
+					defaultActions.separator(),
+					defaultActions.searchWithGoogle({}),
+					defaultActions.cut({}),
+					defaultActions.copy({}),
+					defaultActions.paste({}),
+					defaultActions.separator(),
+					defaultActions.saveImage({}),
+					defaultActions.saveImageAs({}),
+					...(params.linkURL.startsWith('file:///') ? [] : [defaultActions.copyLink({})]),
+					defaultActions.copyImage({}),
+				];
+			}
+		});
+	})();
 }
 
 if (!app.requestSingleInstanceLock()) {
